requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
sentence-transformers
scikit-learn
numpy